fix: abstracting showcase with react-dom specified to fix native bundle for dextrose
