+++ package.json
@@ -118,7 +118,7 @@
     "cz-conventional-changelog": "2.1.0",
     "dashify": "0.2.2",
     "detox": "5.8.2",
-    "dextrose": "3.0.2",
+    "dextrose": "3.0.3",
     "eslint": "4.9.0",
     "eslint-config-airbnb": "16.1.0",
     "eslint-config-prettier": "2.6.0",
