+++ packages/image/image.showcase.js
@@ -1,7 +1,6 @@
 /* eslint-disable react/no-danger */
 import React from "react";
 import { View } from "react-native";
-import ReactDOMServer from "react-dom/server";
 import Image from "./src/image";
 
 const squareUri =
@@ -116,45 +115,6 @@ export default {
           />
         </View>
       )
-    },
-    {
-      type: "story",
-      name: "Server side rendered Image (web only)",
-      platform: "web",
-      component: () => {
-        const markup = {
-          __html: ReactDOMServer.renderToStaticMarkup(
-            <View>
-              <SquareImage
-                style={{
-                  width: 200,
-                  height: 200,
-                  borderRadius: 100,
-                  overflow: "hidden"
-                }}
-              />
-              <SquareImage
-                style={{
-                  width: 100,
-                  height: 100,
-                  borderRadius: 50,
-                  overflow: "hidden"
-                }}
-              />
-              <SquareImage
-                style={{
-                  width: 50,
-                  height: 50,
-                  borderRadius: 25,
-                  overflow: "hidden"
-                }}
-              />
-            </View>
-          )
-        };
-
-        return <div dangerouslySetInnerHTML={markup} />;
-      }
     }
   ]
 };
